Rename _link.py -> device_resident.py
