+++ chainer/device_resident.py
